requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-openai
langchain-ollama
chromadb
faiss-cpu
sentence_transformers
langchain_huggingface
langchain_chroma
langchain_groq
langchain-classic
streamlit
langsmith
fastapi
uvicorn
langserve
sse_starlette